test: Move non-port datetime tests to inline parser tests
The `asciidoctor_rb/document_test.rs` module tracks direct ports of
Asciidoctor's Ruby `document_test.rb`. The time-dependent-attribute tests that
are not ports — the injectable clock (with_reference_time / with_input_mtime)
and resolution during a parse (a `{docdate}` reference, an `ifdef::docdate[]`
directive, and the explicit / value-less / unset doctime sibling cases) —
move into an inline `datetime_attributes` submodule under `parser.rs`'s tests,
next to the code they exercise. The six direct ports stay in
`document_test.rs`. No behavior or coverage change.
